User Variable Voxel Tycoon Directory
Thanks to the contribution of @xmnovotny , the voxel tycoon path has been converted to a User Variable, allowing people to install Voxel Tycoon anywhere on their computer and still be able to contribute to this repo.

An initial description of how to add the 'User Variable' has been provided in the README. Pictures should still be added when someone with an English Windows version can take screenshots of the process (I doubt you want it in Dutch).

Co-Authored-By: xmnovotny <3328134+xmnovotny@users.noreply.github.com>